Fix compilation warning.
int -> intptr_t for better 64 bits portability.

git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@19255 32a0c1e6-9308-0410-80dd-f9d2257a62c7

+++ src/Libs/hyperplane/symbolic_tiling.c
@@ -205,7 +205,7 @@ generate_tile:;
 /* checks if sloop is a perfectly nested loop of depth @depth */
 static bool symbolic_tiling_valid_p(statement sloop, size_t depth)
 {
-    int l;
+    intptr_t l;
     if(depth == 0 ) return true;
     else {
         if(statement_loop_p(sloop) &&
